ADD CPP03/EX00:
This is a combination of 3 commits.
 This is the 1st commit message:

WIP EX00 CPP04:

The makeSound functions are correctly called depending on the type.
Now needs to make the classes canonical ; to implement WrongAnimal and
its derived classes and fill the main with tests.

 This is the commit message #2:

WIP EX00 CPP04:

Canonical forms OK, now need to improve tests and implement Wrong*
classes

 This is the commit message #3:

ADD CPP03/EX00:

Tests are complete