FIX CPP05/EX00 UNSIGNED INT AS CONSTRUCTOR PARAM:
This is a combination of 2 commits.
 This is the 1st commit message:

WIP CPP05/EX00:

This is a combination of 2 commits.
 This is the 1st commit message:

WIP CPP05/EX00:

For some reason, does not accept = overload ??
Good Luck Tommorow Elena

 This is the commit message #2:

WIP CPP05/EX00:

_name is a constant and the _name assignation in the operator overload
is removed.

The sub-classes exception are implemented and the instantiations are
embedded with try catch

 This is the commit message #2:

FIX CPP05/EX00 UNSIGNED INT AS CONSTRUCTOR PARAM:

The grade is saved as unsigned but is passed to the constructor
as an int to prevent conversion and be accurately caught