FIX CPP04 INHERITANCE MALPRACTICE:
Dog and Cat classes do not need to have their own type attribute,
since it is inherited from Animal.

Brain allocation is made smarter and without leaks, without having
to free a brain to reallocate it in assignation operator overload.

The use of Animal copy constructor makes rendundant the assignation
from the source in the derived classes copy constructor

Function getType is inherited from Animal and return Animal's type
attribute so it is redundant in the derived classes and therefore
deleted from them

strings (ideas and idea) are sent as const & from one function to
the other to prevent the use of copies, improve the use of ressources,
improve speed and for good practice