ADD CPP05/EX02 ROBOTOMY FEATURE:
This is a combination of 4 commits.
 This is the 1st commit message:

WIP CPP05/EX02:

Initial commit to create files

 This is the commit message #2:

WIP CPP05/EX02:

AForm has a function execute which is purely virtual and is redefined in
each derived class.

Shrubbery Creation is the only functional one, printing a bonzai
in an outfile

 This is the commit message #3:

WIP CPP05/EX02 :

Robotomy does not randomly pick a message

 This is the commit message #4:

ADD CPP05/EX02 ROBOTOMY FEATURE:

The Robotomy feature is added and generates the success message at
random using srand() at the begining of the function to reinitialise
the seed to generate a new number