WIP EX00/CPP09:
Refacto : function extractData() is created, to prevent redundency;
the date is returned and assigned the correct variable, while the
value is a reference to either the rate or the amount of bitcoins
depending on the calling function. So proud of me to have found this
solution \o/ Pat on the back, Pat on the back.